% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_r2_contribution}
\alias{.extract_r2_contribution}
\title{Extract R² contribution of a single predictor from a ranger model}
\usage{
.extract_r2_contribution(model, predictor_name)
}
\description{
Extract R² contribution of a single predictor from a ranger model
}
\keyword{internal}
