% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R, R/utils.R
\name{.extract_split_density_cached}
\alias{.extract_split_density_cached}
\title{Extract split density for one predictor using a pre-cached treeInfo tibble}
\usage{
.extract_split_density_cached(treeinfo_cache, x_vec, predictor_name, num_bins)

.extract_split_density_cached(treeinfo_cache, x_vec, predictor_name, num_bins)
}
\arguments{
\item{treeinfo_cache}{tibble returned by ranger::treeInfo(), pre-computed}

\item{x_vec}{numeric vector of env values for this predictor}

\item{predictor_name}{character}

\item{num_bins}{integer}
}
\description{
Extract split density for one predictor using a pre-cached treeInfo tibble

Extract split density for one predictor using a pre-cached treeInfo tibble
}
\keyword{internal}
