% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_gf_result}
\alias{new_gf_result}
\title{Constructor for gf_result}
\usage{
new_gf_result(turnover_full, vars, n_models)
}
\arguments{
\item{turnover_full}{long tibble: SNP_ID x variable x x x F_x x importance x r_squared}

\item{vars}{character vector of variables used}

\item{n_models}{number of models processed}
}
\description{
Constructor for gf_result
}
\keyword{internal}
