% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R, R/utils.R
\name{.get_ranger}
\alias{.get_ranger}
\title{Safely extract the ranger model from a list entry (or return as-is)}
\usage{
.get_ranger(entry)

.get_ranger(entry)
}
\description{
Safely extract the ranger model from a list entry (or return as-is)

Safely extract the ranger model from a list entry (or return as-is)
}
\keyword{internal}
