% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.turnover_curve_one}
\alias{.turnover_curve_one}
\title{Compute normalised cumulative turnover curve for one SNP × predictor}
\usage{
.turnover_curve_one(model, treeinfo_cache, x_vec, predictor_name, num_bins)
}
\description{
Compute normalised cumulative turnover curve for one SNP × predictor
}
\keyword{internal}
