% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_models}
\alias{.validate_models}
\title{Validate that models is a named list of ranger objects (or wrapped)}
\usage{
.validate_models(models)
}
\description{
Validate that models is a named list of ranger objects (or wrapped)
}
\keyword{internal}
