% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/todo.R
\name{new_gf_summary}
\alias{new_gf_summary}
\title{Constructor for gf_summary}
\usage{
new_gf_summary(
  snp_summary,
  var_importance,
  var_snp_matrix,
  turnover_concentration,
  curve_cor,
  curve_xy,
  snps_keep,
  zone_counts,
  mean_curve
)

new_gf_summary(
  snp_summary,
  var_importance,
  var_snp_matrix,
  turnover_concentration,
  curve_cor,
  curve_xy,
  snps_keep,
  zone_counts,
  mean_curve
)
}
\description{
Constructor for gf_summary

Constructor for gf_summary
}
\keyword{internal}
